% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_action_button.R
\name{material_action_buttonInput}
\alias{material_action_buttonInput}
\title{<Add Title>}
\usage{
material_action_buttonInput(
  inputId,
  label,
  variant = TRUE,
  outline = TRUE,
  color = c("default", "primary", "secondary"),
  size = c("medium", "small", "large"),
  disabled = FALSE
)
}
\description{
<Add Description>
}
