% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_calendar_keyboard.R
\name{material_calendar_keyboardInput}
\alias{material_calendar_keyboardInput}
\title{<Add Title>}
\usage{
material_calendar_keyboardInput(
  inputId,
  value = NULL,
  label = NULL,
  minDate = NULL,
  maxDate = NULL,
  format = "default",
  disableToolbar = FALSE,
  view = c("year-day", "year-month-day", "year", "year-month"),
  helperText = NULL,
  variant = c("default", "static", "inline"),
  autoOk = TRUE,
  openTo = c("day", "month", "year"),
  animateYearScrolling = TRUE,
  disableFuture = FALSE,
  horizontal_orientation = FALSE,
  outlined = FALSE
)
}
\description{
<Add Description>
}
