% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_calendar.R
\name{material_calendarInput}
\alias{material_calendarInput}
\title{<Add Title>}
\usage{
material_calendarInput(
  inputId,
  value = NULL,
  label = NULL,
  minDate = NULL,
  maxDate = NULL,
  format = "default",
  primary_color = "#353535",
  secondary_color = "#ff9900",
  disableToolbar = FALSE,
  view = c("year-day", "year-month-day", "year", "year-month"),
  helperText = NULL,
  variant = c("default", "static", "inline"),
  autoOk = TRUE,
  openTo = c("day", "month", "year"),
  animateYearScrolling = TRUE,
  disableFuture = FALSE,
  horizontal_orientation = FALSE,
  outlined = FALSE
)
}
\arguments{
\item{value}{format \code{'ymd'}}

\item{minDate}{format \code{'ymd'}}

\item{maxDate}{format \code{'ymd'}}

\item{format}{e.g. \code{'dd/MM/yyyy'}, \code{'yyyy/MM/dd'}, \code{'MM/dd/yyyy'} etc.}
}
\description{
<Add Description>
}
