% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_datetimeInput.R
\name{material_datetimeInputInput}
\alias{material_datetimeInputInput}
\title{<Add Title>}
\usage{
material_datetimeInputInput(
  inputId,
  value = NULL,
  label = NULL,
  minDate = NULL,
  maxDate = NULL,
  primary_color = "#353535",
  secondary_color = "#ff9900",
  autoOk = TRUE,
  keyboardControl = FALSE,
  outlined = FALSE,
  ampm = FALSE,
  disableFuture = FALSE,
  disablePast = FALSE,
  helperText = NULL,
  hideTabs = FALSE,
  variant = c("default", "static", "inline")
)
}
\description{
<Add Description>
}
