% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_slider.R
\name{material_sliderInput}
\alias{material_sliderInput}
\title{<Add Title>}
\usage{
material_sliderInput(
  inputId,
  value,
  default = value,
  step,
  min,
  max,
  track = TRUE,
  disabled = FALSE,
  vertical = FALSE,
  aria_labelled_by = c("continuous-slider", "discrete-slider-small-steps",
    "discrete-slider", "track-false-range-slider"),
  marks = FALSE
)
}
\description{
<Add Description>
}
