% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_switch_button.R
\name{material_switch_buttonInput}
\alias{material_switch_buttonInput}
\title{<Add Title>}
\usage{
material_switch_buttonInput(
  inputId,
  label = NULL,
  size = c("normal", "small"),
  color = c("default", "primary", "secondary"),
  checked = FALSE,
  disabled = FALSE,
  labelPlacement = c("end", "start", "top", "bottom")
)
}
\description{
<Add Description>
}
