% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_timepicker.R
\name{material_timepickerInput}
\alias{material_timepickerInput}
\title{<Add Title>}
\usage{
material_timepickerInput(
  inputId,
  value = NULL,
  label = NULL,
  primary_color = "#353535",
  secondary_color = "#ff9900",
  ampm = FALSE,
  autoOk = TRUE,
  variant = c("default", "static", "inline"),
  showTodayButton = TRUE,
  todayLabel = "now",
  horizontal_orientation = FALSE,
  minutesStep = 1,
  views = c("hm", "hms", "ms"),
  openTo = c("hours", "minutes", "seconds"),
  outlined = FALSE
)
}
\description{
<Add Description>
}
