% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_datetimeInput.R
\name{updateMaterial_datetimeInputInput}
\alias{updateMaterial_datetimeInputInput}
\title{<Add Title>}
\usage{
updateMaterial_datetimeInputInput(
  session,
  inputId,
  value,
  configuration = NULL
)
}
\description{
<Add Description>
}
