% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_slider.R
\name{updateMaterial_sliderInput}
\alias{updateMaterial_sliderInput}
\title{<Add Title>}
\usage{
updateMaterial_sliderInput(session, inputId, value, configuration = NULL)
}
\description{
<Add Description>
}
