% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_switch_button.R
\name{updateMaterial_switch_buttonInput}
\alias{updateMaterial_switch_buttonInput}
\title{<Add Title>}
\usage{
updateMaterial_switch_buttonInput(
  session,
  inputId,
  value,
  disabled,
  checked,
  configuration = NULL
)
}
\description{
<Add Description>
}
